moved entity manager
